architecture cleanup
