+++ src/game/gameState.ts
@@ -1,0 +1,12 @@
+export const GameState = {
+  EXPLORING: 'EXPLORING',
+  IN_MINIGAME: 'IN_MINIGAME',
+  LOADING: 'LOADING'
+} as const;
+
+export type GameStateValue = (typeof GameState)[keyof typeof GameState];
+
+export interface AppState {
+  status: GameStateValue;
+  activeMiniGameId: string | null;
+}
